5_testing/unit_test_example_1/requirements.txt
biopython==1.86
coverage==7.13.5
numpy==2.4.3
